+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Added support for JSONL as input data format
+- Added toast messages extension for displaying non-intrusive notifications in the app via `toastMessage(config)` action
 - Added `markdown` (with `md` alias) text view
 - Added `render` view for text renderer
 - Added `ViewRenderer#getViewsMetadata()` method